Implement all_as_args function
